Added async resolver and changed the API for the resolver
- Third argument changed to an optional option object
- The option object supports defining a custom resolver
- Custom resolvers will get an additional argument, the identifier
